updated cpp files

+++ cpp/src/orchestrator.cpp
@@ -4,13 +4,17 @@
  */
 
 #include "orchestrator.hpp"
+#include "simulation_prestep.hpp"
+#include "predictor.hpp"
+#include "pressure_poisson_solver.hpp"
+#include "corrector.hpp"
 #include "grid_math.hpp"
 #include <stdexcept>
 
 namespace navier_stokes_solver {
 
 NavierStokesOrchestrator::NavierStokesOrchestrator(const GridDimensions& dims, const SolverConfig& config)
-    : dims_(dims), config_(config), total_cells_(dims.nx * dims.ny * dims.nz) {
+    : dims_(dims), config_(config), total_cells_(static_cast<size_t>(dims.nx) * dims.ny * dims.nz) {
     u_star_.resize(total_cells_, 0.0);
     v_star_.resize(total_cells_, 0.0);
     w_star_.resize(total_cells_, 0.0);
